Refactoring Room, changing from interface to an abstract class, for get less code inside each room class!
