+++ src/mindswap/academy/sims/houses/rooms/Living.java
@@ -3,15 +3,9 @@ package mindswap.academy.sims.houses.rooms;
 import mindswap.academy.sims.messages.Messages;
 import mindswap.academy.sims.player.SimsChar;
 
-public class Living implements Room {
+public class Living extends Room {
 
-    private int maxLevelOfCleanliness;
-    private int levelOfCleanliness;
 
-    public Living() {
-        maxLevelOfCleanliness = 100;
-        levelOfCleanliness = 100;
-    }
     @Override
     public void execute(SimsChar simsChar) {
         if(simsChar.getEnergyLevel() >= 20){
@@ -28,23 +22,4 @@ public class Living implements Room {
         simsChar.getPh().sendMessage(Messages.YOU_NEED_TO_REST);
     }
 
-    @Override
-    public void setLevelOfCleanliness(int cleanliness) {
-        levelOfCleanliness = cleanliness;
-    }
-
-    @Override
-    public int getLevelOfCleanliness() {
-        return levelOfCleanliness;
-    }
-
-    @Override
-    public void decreaseCleanness(int levelOfDirt) {
-        levelOfCleanliness -= levelOfDirt;
-    }
-
-    @Override
-    public void cleanRoom() {
-        setLevelOfCleanliness(maxLevelOfCleanliness);
-    }
 }
